back/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
#requests
pymongo[srv]
python-dotenv